Implement support for NeoForge (#13)
* Fix compilation on Java versions other than 8

* Add basic support for NeoForge

This uses Forge's built-in fml.forgeGroup flag to distinguish Forge and NeoForge's package names. Consequently, this flag's value needs to be passed to most installer-related functions.

* Update Javadocs with Forge group parameter

* Add Forge group support for manual installs

* Assume we're using NeoForge if `--fml.forgeGroup` isn't specified

I was hoping not to hardcode any package names here, but this is the only way to unbreak manual installs without unintuitively asking users to add the flag themselves.